pyproject.toml
[project]
name = "update-mypy-hook"
version = "0.1.0"
description = "Update the additional dependencies of mypy pre-commit hook with uv.lock"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
    "typing-extensions; python_version < '3.11'"
]

[project.scripts]
update-mypy-hook = "update_mypy_hook:main"

[build-system]
requires = ["setuptools>=75.6.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
lint.select = [
    # "E", # pycodestyle errors
    # "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    # "C", # flake8-comprehensions
    # "B", # flake8-bugbear
]

[tool.uv]
default-groups = ["dev", "mypy"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.9",
]
mypy = [
    "mypy>=1.15.0",
    "types-pyyaml>=6.0.12.20250402",
]
